requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.0.0.post2
pytest==7.3.1
httpx==0.24.0
alembic==1.11.2
psycopg2-binary==2.9.6
